requirements.txt
# ─── Web Framework ───────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets>=12.0

# ─── LangGraph + LangChain ───────────────────────────────────
langgraph==0.2.14
langchain==0.2.12
langchain-core==0.2.29
langchain-google-genai==1.0.8

# ─── Observability ───────────────────────────────────────────
langfuse==2.36.1

# ─── Database ────────────────────────────────────────────────
sqlalchemy==2.0.31
aiosqlite==0.20.0

# ─── File Parsing ────────────────────────────────────────────
PyMuPDF==1.24.7
python-docx==1.1.2
aiofiles==23.2.1

# ─── Job Scraping ────────────────────────────────────────────
apify-client==1.7.1

# ─── Gmail API ───────────────────────────────────────────────
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.139.0

# ─── Utilities ───────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0
httpx==0.27.0